TEst input entered

+++ src/main/java/htl/steyr/computerRent/model/Device.java
@@ -50,7 +50,7 @@ public class Device {
         this.modelName = modelName;
     }
 
-    public float getPrice() {
+    public int getPrice() {
         return price;
     }
